added initial tests
