+++ .bundle/config
@@ -1,0 +1,2 @@
+---
+BUNDLE_BIN: "bin"
