+++ Gemfile
@@ -1,0 +1,16 @@
+source 'https://rubygems.org'
+
+ruby RUBY_VERSION
+
+gem 'rubocop'
+
+group :test do
+  gem 'capybara'
+  gem 'chromedriver-helper'
+  gem 'colorize'
+  gem 'pry'
+  gem 'rack-jekyll'
+  gem 'rb-readline'
+  gem 'rspec'
+  gem 'selenium-webdriver'
+end
