+++ circle.yml
@@ -1,0 +1,12 @@
+machine:
+  ruby:
+    version:
+      2.3.1
+
+dependencies:
+  pre:
+    - bundle install --binstubs
+
+test:
+  post:
+    - bundle exec rspec
