+++ lib/jekyll_frontmatter_tests/jekyll_frontmatter_tests_config.rb
@@ -1,0 +1,15 @@
+require 'pry'
+require 'yaml'
+require 'jekyll'
+
+class FrontmatterTests < Jekyll::Command
+  class << self
+    def schema_config
+      config = Jekyll.configuration
+      unless config.key?('frontmatter_tests')
+        config['frontmatter_tests'] = { 'path' => File.join('deploy', 'tests', 'schema') }
+      end
+      config['frontmatter_tests']
+    end
+  end
+end
