+++ spec/support/tests/schema/_tests.yml
@@ -1,0 +1,19 @@
+---
+name: "String"
+first_name: "String"
+last_name: "String"
+full_name: "String"
+role: "String"
+city: "String"
+state: "String"
+role: "String"
+published: Boolean
+config:
+  path: '_authors'
+  ignore:
+  - README.md
+  - ..
+  - .
+  - .DS_Store
+  optional: ['tumblr_url', 'github', 'twitter', 'role', 'city', 'state', 'team', 'last_name', 'team']
+---
