+++ test/jekyll_project/.gitignore
@@ -1,0 +1,4 @@
+_site
+.sass-cache
+.jekyll-metadata
+bin
