+++ test/jekyll_project/.rspec
@@ -1,0 +1,2 @@
+--color
+--require spec_helper
