+++ test/jekyll_project/_authors/brian.md
@@ -1,0 +1,9 @@
+---
+name: Brian
+favorite_ice_cream:
+  - vanilla
+  - "cookies 'n cream"
+  - mint chocolate chip
+---
+
+Hi, my name is Brian
