+++ tests/schema/tags.yml
@@ -1,0 +1,4 @@
+- tag
+- tag2
+- tag3
+- tag4
